third_party/blink/web_tests/fast/forms/select-popup/popup-menu-event-order-expected.txt

PASS internals.pagePopupWindow is null
PASS menu.value is "garply"
PASS internals.selectMenuListText(menu) is "garply"
PASS events[0] is "change"
PASS events[1] is "mouseup"
PASS events[2] is "click"
PASS successfullyParsed is true

TEST COMPLETE

foo
bar
baz
garply
